// src/libs/order/order-validation-request.interface.ts
export default interface OrderValidationRequestInterface {
  tanggal_order: any;
  pelanggan_hp: any;
  pelanggan_nama: any;
  pelanggan_alamat: any;
  pelanggan_wilayah: any;
  pelanggan_kota: any;
  pelanggan_provinsi: any;
  kode_pos: any;
  metode_pembayaran: any;
  nilai_cod: any;
  nilai_barang: any;
  nama_paket: any;
  berat: any;
  jumlah: any;
  catatan_pengiriman: any;
  remark_1: any;
  remark_2: any;
  remark_3: any;
}